try to fix

+++ math/0x03-probability/poisson.py
@@ -9,10 +9,10 @@ class Poisson:
         if data is None:
             if (lambtha <= 0):
                 raise ValueError("lambtha must be a positive value")
-            self.lambtha = lambtha
-        else:    
-            if isinstance (data, list) is False:
+            self.lambtha = float(lambtha)
+        else:
+            if isinstance(data, list) is False:
                 raise TypeError("data must be a list")
             if(len(data)) < 1:
                 raise ValueError("data must contain multiple values")
-            self.lambtha = float(sum(data) / len(data))
+            self.lambtha = (sum(data) / len(data))
